Add organisation to grades, move resource to /grades.
