+++ src/main/java/uk/gov/cshr/civilservant/api/OrganisationController.java
@@ -1,15 +1,12 @@
 package uk.gov.cshr.civilservant.api;
 
-import com.google.common.collect.Iterables;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
-import uk.gov.cshr.civilservant.domain.Grade;
 import uk.gov.cshr.civilservant.domain.Organisation;
-import uk.gov.cshr.civilservant.repository.GradeRepository;
 import uk.gov.cshr.civilservant.repository.OrganisationRepository;
 
 import java.util.Optional;
@@ -26,15 +23,11 @@ public class OrganisationController {
 
     private static final Logger LOGGER = LoggerFactory.getLogger(OrganisationController.class);
 
-    private GradeRepository gradeRepository;
-
     private OrganisationRepository organisationRepository;
 
     @Autowired
-    public OrganisationController(GradeRepository gradeRepository, OrganisationRepository organisationRepository) {
-        checkArgument(gradeRepository != null);
+    public OrganisationController(OrganisationRepository organisationRepository) {
         checkArgument(organisationRepository != null);
-        this.gradeRepository = gradeRepository;
         this.organisationRepository = organisationRepository;
     }
 
@@ -58,23 +51,5 @@ public class OrganisationController {
         return result
                 .map(organisation -> new ResponseEntity<>(new OrganisationResource(organisation, false), OK))
                 .orElseGet(() -> new ResponseEntity<>(NOT_FOUND));
-    }
-
-    @GetMapping("/{organisationId}/grades")
-    @Transactional(readOnly = true)
-    public ResponseEntity<Results<GradeResource>> listGrades(@PathVariable Long organisationId) {
-        LOGGER.debug("Listing grades for organisation with id {}", organisationId);
-
-        Optional<Organisation> result = organisationRepository.findById(organisationId);
-
-        return result
-                .map(organisation -> {
-                    Iterable<Grade> grades = organisation.getGrades();
-                    if (Iterables.isEmpty(grades)) {
-                        grades = gradeRepository.findByCoreTrue();
-                    }
-                    return ResponseEntity.ok(new Results<>(StreamSupport.stream(grades.spliterator(), false).map(GradeResource::new).collect(toList())));
-                })
-                .orElseGet(() -> ResponseEntity.notFound().build());
     }
 }
