+++ src/main/java/uk/gov/cshr/civilservant/repository/GradeRepository.java
@@ -7,5 +7,5 @@ import uk.gov.cshr.civilservant.domain.Grade;
 @Repository
 public interface GradeRepository extends CrudRepository<Grade, Long> {
 
-    Iterable<Grade> findByCoreTrue();
+    Iterable<Grade> findByOrganisationIsNull();
 }
