+++ src/test/java/uk/gov/cshr/civilservant/api/GradeControllerTest.java
@@ -10,11 +10,15 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import uk.gov.cshr.civilservant.domain.Grade;
 import uk.gov.cshr.civilservant.domain.Organisation;
 import uk.gov.cshr.civilservant.repository.GradeRepository;
+import uk.gov.cshr.civilservant.repository.OrganisationRepository;
 
 import java.util.Optional;
 
+import static com.google.common.collect.Lists.newArrayList;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.hasSize;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
@@ -32,10 +36,76 @@ public class GradeControllerTest {
     @Mock
     private GradeRepository gradeRepository;
 
+    @Mock
+    private OrganisationRepository organisationRepository;
+
     @Before
     public void setup() {
         initMocks(this);
         mockMvc = standaloneSetup(controller).build();
+    }
+
+    @Test
+    public void shouldReturnEmptyListForGradesQueryOfUnknownOrganisation() throws Exception {
+
+        final Long id = 1L;
+
+        when(organisationRepository.findById(id)).thenReturn(Optional.empty());
+
+        mockMvc.perform(
+                MockMvcRequestBuilders.get("/grades?organisation={id}", id)
+                        .accept(MediaType.APPLICATION_JSON))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.results", hasSize(0)));
+    }
+
+    @Test
+    public void shouldReturnOrganisationsGradesForGradesQueryOfOrganisationWithGrades() throws Exception {
+
+        final Long id = 1L;
+        final String gradeCode = "gradeCode";
+        final String gradeName = "gradeName";
+
+        Organisation organisation = new Organisation("code", "name");
+        organisation.addGrade(new Grade(gradeCode, gradeName));
+
+        when(organisationRepository.findById(id)).thenReturn(Optional.of(organisation));
+
+        mockMvc.perform(
+                MockMvcRequestBuilders.get("/grades?organisation={id}", id)
+                        .accept(MediaType.APPLICATION_JSON))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.results", hasSize(1)))
+                .andExpect(jsonPath("$.results[0].code", equalTo(gradeCode)))
+                .andExpect(jsonPath("$.results[0].name", equalTo(gradeName)));
+
+    }
+
+    @Test
+    public void shouldReturnDefaultGradesForGradesQueryOfOrganisationWithNoGrades() throws Exception {
+
+        final Long id = 1L;
+        final String gradeCode = "gradeCode";
+        final String gradeName = "gradeName";
+
+        Organisation organisation = new Organisation("code", "name");
+        Grade grade = new Grade(gradeCode, gradeName);
+
+        when(organisationRepository.findById(id)).thenReturn(Optional.of(organisation));
+        when(gradeRepository.findByOrganisationIsNull()).thenReturn(newArrayList(grade));
+
+        mockMvc.perform(
+                MockMvcRequestBuilders.get("/grades?organisation={id}", id)
+                        .accept(MediaType.APPLICATION_JSON))
+                .andDo(print())
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.results", hasSize(1)))
+                .andExpect(jsonPath("$.results[0].code", equalTo(gradeCode)))
+                .andExpect(jsonPath("$.results[0].name", equalTo(gradeName)));
+
+        verify(gradeRepository, times(1)).findByOrganisationIsNull();
     }
 
     @Test
